Step 5 for getting updated porduct price if big than 1 minute confirm order.
Step 5 for getting Updated Porduct price if big than 1 minute confirm
order.

Modify the app to save the information about the product price when
creating an order so
it’s not dependent on future price changes.
Assume that a price is valid for 1 minute, if you won’t confirm it in
that timespan the prices get updated to the current ones.